# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Danshin Genealogy",
        "description": "Доступ ко всем данным только через API токен,\n        в противном случае будет возвращаться ошибка 401.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/marriage/roles/{gender}": {
            "get": {
                "operationId": "1d492c94a9d97e8c3aa459d22467f1cd",
                "parameters": [
                    {
                        "name": "gender",
                        "in": null,
                        "description": "Id gender.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список допустимых ролей (выборка по полу) в браке \n                    или близких отношениях между лицами."
                    },
                    "404": {
                        "description": "Id запрошенного пола не найдено."
                    }
                }
            }
        },
        "/api/marriage/possible/{personId}": {
            "get": {
                "operationId": "315a9bc1a95a3b77c55d77940a1160a1",
                "parameters": [
                    {
                        "name": "person_id",
                        "in": null,
                        "description": "Id запрашиваемого лица.",
                        "required": false
                    },
                    {
                        "name": "birth_date",
                        "in": null,
                        "description": "Дата рождения запрашиваемого лица.",
                        "required": false
                    },
                    {
                        "name": "role_id",
                        "in": null,
                        "description": "Id роли для родителя.",
                        "required": true
                    },
                    {
                        "name": "mariages",
                        "in": null,
                        "description": "Массив id лиц, состоящих в отношениях, с запрашиваемым лицом.",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список допустимых лиц для брака или сожительства."
                    },
                    "422": {
                        "description": "Неверные параметры запроса."
                    }
                }
            }
        },
...
        "/api/parent/possible/{personId}/{roleParent}": {
            "get": {
                "operationId": "143ef865cb00bd4039ac0472bb29e8dd",
                "parameters": [
                    {
                        "name": "person_id",
                        "in": null,
                        "description": "Id запрашиваемого лица.",
                        "required": false
                    },
                    {
                        "name": "birth_date",
                        "in": null,
                        "description": "Дата рождения запрашиваемого лица.",
                        "required": false
                    },
                    {
                        "name": "role_id",
                        "in": null,
                        "description": "Id роли для родителя.",
                        "required": true
                    },
                    {
                        "name": "mariages",
                        "in": null,
                        "description": "Массив id лиц, состоящих в отношениях, с запрашиваемым лицом.",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список допустимых родителей для запрошенного лица."
                    },
                    "422": {
                        "description": "Неверные параметры запроса."
                    }
                }
            }
        },
...
    }
}